+++ app/javascript/packs/dashboard.jsx
@@ -2,39 +2,46 @@
 // like app/views/layouts/application.html.erb. All it does is render <div>Hello React</div> at the bottom
 // of the page.
 import 'bootstrap/dist/css/bootstrap.min.css';
+import './dashboard.css'
 
-import React from 'react';
+import React, {useState, useEffect} from 'react';
 import ReactDOM from 'react-dom';
+import { getUser } from './api';
 import PropTypes from 'prop-types';
 
 import {Container, Col, Row, Nav} from 'react-bootstrap';
 
-const Hello = props => (
-  <Container fluid={true}>
-    <Row>
-      <Col xs={1}>
-        <Nav defaultActiveKey="/home" className="flex-column">
-          <Nav.Link href="/home">Dashboard</Nav.Link>
-        </Nav>
-      </Col>
-      <Col xs={11}>
-        Hello {props.name}!
-      </Col>
-    </Row>
-  </Container>
-)
+import {EditActivity} from './editActivity'
 
-Hello.defaultProps = {
-  name: 'David'
-}
+const Dashboard = (props) => {
+  const [user, setUser] = useState('');
 
-Hello.propTypes = {
-  name: PropTypes.string
+  useEffect(() => {
+    setUser('Lucas');
+  }, []);
+
+  return (
+    <Container fluid={true}>
+      <Row>
+        <Col className="menu">
+          <Nav defaultActiveKey="/home" className="flex-column">
+            <Nav.Link href="/home">Dashboard</Nav.Link>
+          </Nav>
+        </Col>
+        <Col xs={12}>
+          <Row>Hello {user}!</Row>
+          <Row>
+            <EditActivity />
+          </Row>
+        </Col>
+      </Row>
+    </Container>
+  )
 }
 
 document.addEventListener('DOMContentLoaded', () => {
   ReactDOM.render(
-    <Hello name="React" />,
+    <Dashboard />,
     document.getElementById('root'),
   )
 })