+++ app/models/activity.rb
@@ -1,3 +1,4 @@
 class Activity < ApplicationRecord
   belongs_to :user
+  has_one_attached :fit_file
 end