+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 # CHANGE LOG
 
+### 1-5-2020
+ - Added basic file upload
+
 ### 1-2-2020
  - Added activity controller and model
  - Added Redux, React Router and, React Bootstrap